Fix a team membership migration bug
The create team membership API fails if an invalid email, phone, or
name are supplied. We can get rid of all of those params because the
we can look up the user by user id and user id should always be
available.

In addition, the create team membership API will fail if an empty string
is passed for the url. We have to use the latest SDK where we can omit
the url param.